Conflicts:
	src/remote_repo_cmds.go

+++ src/remote_builds_cmd.go
@@ -1,3 +1,5 @@
+// +build TODO
+
 package src
 
 import (
@@ -49,7 +51,7 @@ func (c *BuildsCmd) Execute(args []string) error {
 		Failed:      c.Failed,
 		Sort:        c.Sort,
 		Direction:   c.Direction,
-		ListOptions: sourcegraph.ListOptions{PerPage: c.N},
+		ListOptions: sourcegraph.ListOptions{PerPage: int32(c.N)},
 	}
 	builds, _, err := cl.Builds.List(opt)
 	if err != nil {
